perf: 主机事件处理性能优化 #1650
重构业务、主机、主机关系事件监听及处理相关代码，补充Trace与Metrics数据。
